@@ -26,7 +26,7 @@ dependencies:
     sdk: flutter
   flutter_animate: ^4.5.0
   animate_do: ^3.3.4
-  serverpod_flutter: 2.2.0
+  serverpod_flutter: 2.3.0
   file_picker: ^5.0.0
   animated_text_kit: ^4.2.2
 
